D3DXVec3Length -> D3DXVec3LengthSq, 그리고 warning 좀 잡음

+++ Pooptube/Client/Tree1.cpp
@@ -76,7 +76,7 @@ void Tree1::_SetBillBoardTexture()
 	float angle = pooptube::CalculateAngle( GetFrontVector(), mBillBoard->GetFrontVector() );
 	angle /= D3DX_PI;
 	angle *= 4;
-	int dir( angle );
+	int dir( static_cast<int>(angle) );
 	dir += 8;
 	dir %= 8;
 	switch( dir ){
